[ filter virtual range scan ] - std::pow gcc bug fixed; - immediate-mode rendering restored in SimpleRenderer.h; - pre-filtering: added an OpenGL capabilities check to ensure that everything is supported; - the looking cone axis is now a RichPoint3f parameter instead of three float parameters; - now the vcg renderer is used to generate the color map.. but textures are not recognized.

+++ src/fgt/filter_virtual_range_scan/vs/utils.h
@@ -213,7 +213,7 @@ namespace vs
             if( f.open( QFile::WriteOnly ) )
             {
                 GLfloat* pix = download( lod );
-                int tmpSide = (int)side / pow( 2.0, lod );
+                int tmpSide = (int)side / (int)( pow( 2.0, lod ) );
                 QTextStream stream( &f );
 
                 char buf[300];
